Precompile common workflows with PrecompileTools
Precompilation can take a bit long, since we specialise on a lot of
things (the choice of kernel, the kernel support size, the number of
simultaneous transforms, the number of dimensions, the element types...).
This means there's a lot of possible combinations, which can make the
precompilation times (and the resulting binary sizes) explode.
So we only precompile some of the possible combinations (e.g. only for
the default kernel).